+++ bika/lims/upgrade/v01_02_009.py
@@ -5,6 +5,10 @@
 # Copyright 2018 by it's authors.
 # Some rights reserved. See LICENSE.rst, CONTRIBUTORS.rst.
 
+import time
+import transaction
+
+from bika.lims import api
 from bika.lims import logger
 from bika.lims.config import PROJECTNAME as product
 from bika.lims.upgrade import upgradestep
@@ -28,6 +32,48 @@ def upgrade(tool):
     logger.info("Upgrading {0}: {1} -> {2}".format(product, ver_from, version))
 
     # -------- ADD YOUR STUFF HERE --------
+    migrate_attachment_report_options(portal)
+
+    # Reindex object security for client contents (see #991)
+    reindex_client_local_owner_permissions(portal)
 
     logger.info("{0} upgraded to version {1}".format(product, version))
     return True
+
+
+def reindex_client_local_owner_permissions(portal):
+    """https://github.com/senaite/senaite.core/issues/957 Reindex bika_setup
+    objects located in clients to give proper permissions to client contacts.
+    """
+    start = time.time()
+    bsc = portal.bika_setup_catalog
+    uids = [c.UID() for c in portal.clients.objectValues()]
+    brains = bsc(getClientUID=uids)
+    total = len(brains)
+    for num, brain in enumerate(brains):
+        ob = brain.getObject()
+        logger.info("Reindexing permission for {}/{} ({})"
+                    .format(num, total, ob.absolute_url()))
+        ob.reindexObjectSecurity()
+    end = time.time()
+    logger.info("Fixing local owner role on client objects took {:.2f}s"
+                .format(end-start))
+    transaction.commit()
+
+    
+def migrate_attachment_report_options(portal):
+    """Migrate Attachments with the report option "a" (attach in report)
+       to the option to "i" (ignore in report)
+    """
+    attachments = api.search({"portal_type": "Attachment"})
+    total = len(attachments)
+    logger.info("Migrating 'Attach to Report' -> 'Ingore in Report' "
+                "for %d attachments" % total)
+    for num, attachment in enumerate(attachments):
+        obj = api.get_object(attachment)
+
+        if obj.getReportOption() in ["a", ""]:
+            obj.setReportOption("i")
+            obj.reindexObject()
+            logger.info("Migrated Attachment %s" % obj.getTextTitle())
+