+++ CHANGES.rst
@@ -7,17 +7,33 @@ Changelog
 **Added**
 
 - #1001 Option to show Interim fields on COA
+- #1024 Function to get the Verifiers from an Analysis Request
+- #1019 Support for min and max warns in range charts
+- #1003 Alphanumeric numbering in sequential IDs generator
 
 **Changed**
 
+- #1012 Refactored Contacts listing
+- #1010 Increased max length of Results options to 255 chars (was 40)
 
 **Removed**
+
+- #1005 Removed databasesanitize package
+- #992 Removed "Attach" report option for Attachments
 
 
 **Fixed**
 
+- #1022 Date Received saved as UTC time
+- #1018 Fix AR Add cleanup after template removal
+- #1014 ReferenceWidget does not handle searches with null/None
+- #1008 Previous results from same batch are always displayed in reports
+- #1013 ARs and Samples from other clients are listed when logged in as contact
+- #991 New client contacts do not have access to their own AR Templates
 - #996 Hide checkbox labels on category expansion
 - #990 Fix client analysisspecs view
+- #899 Secondary AR changes received date of sample
+- #888 Order of Interim Fields not maintained on ARs
 
 
 **Security**